created a compoentn wrapper to include the sidenav in all guarded components
