+++ src/components/sideNav/SideNav.css
@@ -3,7 +3,7 @@
   width: 8.5%;
   min-width: 100px;
   height: 100vh;
-  position: absolute;
+  /* position: absolute; */
   left: 0px;
   top: 0px;
   background: var(--blue);
